pyproject.toml
[tool.poetry]
name = "memo-sql-bot"
version = "1"
description = "A memo discord.py SQLite bot"
authors = ["timelessnesses <mooping3roblox@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
aiodns = "3.0.0"
aiohttp = "3.7.4.post0"
aiosignal = "1.2.0"
async-timeout = "3.0.1"
attrs = "21.4.0"
Brotli = "1.0.9"
brotlipy = "0.7.0"
cchardet = "2.1.7"
cffi = "1.15.0"
chardet = "4.0.0"
charset-normalizer = "2.0.12"
frozenlist = "1.3.0"
idna = "3.3"
multidict = "6.0.2"
orjson = "3.6.7"
pycares = "4.1.2"
pycparser = "2.21"
python-dotenv = "0.19.2"
typing-extensions = "4.1.1"
yarl = "1.7.2"
uvloop = {version = "^0.16.0", platform = "linux"}
"discord.py" = {git = "https://github.com/Rapptz/discord.py"}
psutil = "^5.9.0"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
